Move composite slo components into folder
