+++ x-pack/solutions/observability/plugins/slo/public/pages/slos/slos.tsx
@@ -25,7 +25,7 @@ import { useLicense } from '../../hooks/use_license';
 import { usePermissions } from '../../hooks/use_permissions';
 import { usePluginContext } from '../../hooks/use_plugin_context';
 import { LoadingPage } from '../loading_page';
-import { CompositeSloList } from './components/composite_slo_list';
+import { CompositeSloList } from './components/composite_slo/composite_slo_list';
 import { CreateSloBtn } from './components/common/create_slo_btn';
 import { SloList } from './components/slo_list';
 import { SloListSearchBar } from './components/slo_list_search_bar';
